Backend/requirements.txt
Flask
Flask-Cors
boto3
botocore
pandas
openpyxl
sdv
python-dotenv
Gunicorn
